Move ColaManga to single source
The other override, MangaDig, for which this multisrc was created, was removed in #5319.

The extension stays broken and the version code is not bumped.
