--- lib-multisrc/colamanga/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("lib-multisrc")
-}
-
-baseVersionCode = 9
-
-dependencies {
-    api(project(":lib:synchrony"))
-}
