+++ src/zh/onemanhua/build.gradle
@@ -1,10 +1,12 @@
 ext {
     extName = 'COLAMANGA'
     extClass = '.Onemanhua'
-    themePkg = 'colamanga'
-    baseUrl = 'https://www.colamanga.com'
-    overrideVersionCode = 15
+    extVersionCode = 24
     isNsfw = true
 }
 
 apply from: "$rootDir/common.gradle"
+
+dependencies {
+    implementation(project(":lib:synchrony"))
+}
